+++ src/parallel_starter.h
@@ -13,21 +13,24 @@ namespace cec {
 
     class unique_models_input {
     public:
-        unique_models_input(const mat &x, vector<unique_ptr<model>> &&models)
+        unique_models_input(const mat &x, vector<unique_ptr<model>> &&models,
+                            vector<double> weights)
                 : x(x),
-                  models(std::move(models)) {}
+                  models(std::move(models)),
+                  weights(std::move(weights)) {}
 
         unique_models_input(unique_models_input &) = delete;
 
         unique_models_input(unique_models_input &&) = default;
 
         clustering_input get() {
-            return clustering_input(x, models);
+            return clustering_input(x, models, weights);
         }
 
     private:
         const mat &x;
         vector<unique_ptr<model>> models;
+        vector<double> weights;
     };
 
     class parallel_starter {
@@ -130,22 +133,25 @@ namespace cec {
         using subtask = mp_start_subtask;
 
         explicit multiple_starts_task(cec_starter::parameters params, const mat &x,
-                                      const vector<shared_ptr<model_spec>> &model_specs)
+                                      const vector<shared_ptr<model_spec>> &model_specs,
+                                      const vector<double> &weights)
                 : params(params),
                   x(x),
-                  model_specs(model_specs) {}
+                  model_specs(model_specs),
+                  weights(weights) {}
 
         subtask operator()(int starts) const {
             unique_ptr<clustering_starter> starter = make_unique<cec_starter>(params);
             return mp_start_subtask(std::move(starter),
                                     vector<unique_ptr<clustering_processor>>(),
-                                    unique_models_input(x, model_spec::create_models(model_specs)), starts);
+                                    unique_models_input(x, model_spec::create_models(model_specs), weights), starts);
         }
 
     private:
         const cec_starter::parameters params;
         const mat &x;
         const vector<shared_ptr<model_spec>> &model_specs;
+        const vector<double> weights;
     };
 
     class start_and_split_task {
@@ -155,12 +161,14 @@ namespace cec {
         explicit start_and_split_task(cec_starter::parameters init_cl_params,
                                       split_starter::parameters split_params,
                                       const mat &x,
-                                      const vector<shared_ptr<model_spec>> &model_specs
+                                      const vector<shared_ptr<model_spec>> &model_specs,
+                                      const vector<double> &weights
         )
                 : init_cl_params(init_cl_params),
                   split_params(split_params),
                   x(x),
-                  model_specs(model_specs) {}
+                  model_specs(model_specs),
+                  weights(weights) {}
 
         subtask operator()(int starts) {
             unique_ptr<clustering_starter> starter = make_unique<cec_starter>(init_cl_params);
@@ -170,7 +178,7 @@ namespace cec {
             return mp_start_subtask(
                     std::move(starter),
                     std::move(cl_procs),
-                    unique_models_input(x, model_spec::create_models(model_specs)),
+                    unique_models_input(x, model_spec::create_models(model_specs), weights),
                     starts);
         }
 
@@ -179,6 +187,7 @@ namespace cec {
         split_starter::parameters split_params;
         const mat &x;
         const vector<shared_ptr<model_spec>> &model_specs;
+        const vector<double> weights;
     };
 }